+++ CHANGELOG.md
@@ -2,8 +2,19 @@
 
 ## master (unreleased)
 
-* Your contribution
 * Added relation helper methods `has_one`, `has_many`, `belongs_to` ([#49](https://github.com/petalmd/bright_serializer/pull/49))
+* Allow to evaluate entity values with a callable lambda. ([#88](https://github.com/petalmd/bright_serializer/pull/88))
+* Fix `FrozenError (can't modify frozen Array)` when parsing entity. ([#83](https://github.com/petalmd/bright_serializer/pull/83))
+* Added the support to use instance methods from a serializer class in the library ([#85](https://github.com/petalmd/bright_serializer/pull/85))
+* Use real coveralls_reborn gem
+
+## 0.2.5 (2021-03-08)
+
+* When serializing an Hash, check present of the key before trying string ([#57](https://github.com/petalmd/bright_serializer/pull/57))
+
+## 0.2.4 (2021-02-19)
+
+* Try symbol and string keys when the object to serialize is an Hash ([#54](https://github.com/petalmd/bright_serializer/pull/54))
 
 ## 0.2.3 (2021-01-04)
 
@@ -26,7 +37,7 @@
 * Define entity in serializer for grape_swagger ([#9](https://github.com/petalmd/bright_serializer/pull/9))
 
 ## 0.1.1 (2020-07-13)
- 
+
 * Add description in gemspec file
 * Add content in CHANGELOG.md